requirements.txt
Flask>=3.0.3

python-dotenv>=1.0.1

pydantic>=2.7.0

requests>=2.32.0

assemblyai>=0.26.0

google-generativeai>=0.7.2

werkzeug>=3.0.2

itsdangerous>=2.2.0

Jinja2>=3.1.4